cast!

+++ qiskit/transpiler/coupling.py
@@ -143,7 +143,7 @@ class CouplingMap:
         lengths = nx.all_pairs_shortest_path_length(self.graph.to_undirected(as_view=True))
         lengths = dict(lengths)
         size = len(lengths)
-        cmap = np.zeros((size, size))
+        cmap = np.zeros((size, size), dtype=int)
         for idx in range(size):
             cmap[idx, np.fromiter(lengths[idx].keys(), dtype=int)] = np.fromiter(
                 lengths[idx].values(), dtype=int)
@@ -168,7 +168,7 @@ class CouplingMap:
             raise CouplingError("%s not in coupling graph" % (physical_qubit2,))
         if self._dist_matrix is None:
             self._compute_distance_matrix()
-        return self._dist_matrix[physical_qubit1, physical_qubit2]
+        return int(self._dist_matrix[physical_qubit1, physical_qubit2])
 
     def shortest_undirected_path(self, physical_qubit1, physical_qubit2):
         """Returns the shortest undirected path between physical_qubit1 and physical_qubit2.
